% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN.R
\name{CAN}
\alias{CAN}
\title{Cross validation-based Aggregation Network (CAN)}
\usage{
CAN(
  method = NULL,
  M,
  y = NULL,
  fraction = 0.5,
  tm = 20,
  K = 100,
  pk = NULL,
  Nmax = 5,
  amin = 0,
  amax = 1,
  maxIter = 500,
  popSize = 200,
  parallel = TRUE,
  monitor = FALSE
)
}
\arguments{
\item{method}{A string speficifying the algorithm to use for EQO. Possible options include \code{"bls_c"} for a continuous variable with BLS, \code{"ga_c"} for a continuous variable with GA and \code{"ga_u"} for a uniform trait with GA.}

\item{M}{A matrix or dataframe of taxa in the microbiome (samples as rows and taxa as columns).}

\item{y}{A vector or matrix of trait. y is optional when pattern is \code{"u"}. y is a vector when pattern is \code{"c"}.}

\item{fraction}{A number (default 0.5) indicating the fraction of samples to be randomly selected in the training set.}

\item{tm}{A number (default 20) indicating the times of cross-validation to perform.}

\item{K}{(for \code{"bls_c"}) A sufficiently large number (default 100) required for linearization by the optimizer.}

\item{pk}{(for \code{"ga_c"},\code{"ga_d"} or \code{"ga_u"}) A binary vector indicating partially known functional group based on a priori knowledge, with 1 for species forced to be included in the targeted group and 0 for the other unknown species. Length of this vector should be equal to the total number of species in the microbiome. EQO will then search on the basis of the provided partial known group without removing the designated species.}

\item{Nmax}{A number (default 10) for regularization, specifying the maximal number of taxa allowed in the final group.}

\item{amin}{A number (default 0) specifying the lower bound of average relative abundance of the final group across all samples (only applied to a uniform trait to avoid trivial solutions).}

\item{amax}{A number (default 1) specifying the upper bound of average relative abundance of the final group across all samples (only applied to a uniform trait to avoid trivial solutions).}

\item{maxIter}{A number (default 500) for GA package, specifying the maximum number of iterations to run before the GA search is halted.}

\item{popSize}{A number (default 200) for GA package, specifying the initial population size for the GA search.}

\item{parallel}{Optional argument for GA package (default TRUE), indicating whether parallel computing is needed. See GA documentation for full details.}

\item{monitor}{Optional argument for GA package (default FALSE). \code{monitor=plot} indicates that algorithm implementation is visualized as a plot. \code{monitor=gaMonitor} indicates that algorithm implementation is printed as text. \code{monitor=FALSE} indicates the output should be suppressed.}
}
\value{
A list with the following components.
\itemize{
\item output - a list with taxa selected in each time of optimization and the corresponding performances in cross-validation
\item nodes - a table of nodes for CAN
\item edges - a table of edges for CAN
}
}
\description{
This function splits the microbiome dataset into a training set and a test set.
Then it generates a group of taxa by Ensemble Quotient Optimization for the training set
and evaluates the performance of the group with the test set. Finally, the performance of
all taxa selected in any group is aggregated and normalized into its relative importance.
}
\examples{
# This may take several minutes.
can = CAN("ga_c",Microbiome,trait,maxIter=100)
}
\seealso{
\link{EQO_ga},
\link{EQO_bls},
}
