% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbiome}
\alias{Microbiome}
\title{Demo data to display mEQO package}
\format{
Three arrays included.
\itemize{
\item Microbiome - A matrix with 20 rows(samples) and 8 columns(species).
\item trait_d - A 20*2 matrix indicates the category of each sample(20 samples in 2 categories).
The target \code{y} in \code{patten} "d".
\item trait - A Vector length 20. The target \code{y} in \code{patten} "c".
}
}
\source{
\link{https://github.com/Xiaoyu2425/mEQO}
}
\description{
A small sample dataset of a biological experiment to display all functions in package "mEQO".
}
\examples{
View(Microbiome)
can<-CAN("ga_c",Microbiome,trait,maxIter=100)
CAN_plot(can)
}
