% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQO_bls.R
\name{EQO_bls}
\alias{EQO_bls}
\title{Ensemble Quotient Optimization with a boolean least square (BLS) algorithm}
\usage{
EQO_bls(M, y, Nmax, K)
}
\arguments{
\item{M}{A matrix of taxa in the microbiome (samples as rows and taxa as columns).}

\item{y}{A vector of a continuous functional trait (length as number of samples).}

\item{Nmax}{A number (default 10) specifying the maximal number of taxa allowed in the final group.}

\item{K}{A sufficiently large number (default 100) required for linearization by the optimizer.}
}
\value{
A list with the following components.
\itemize{
\item x - A binary numeric vector specifying presence/absence of taxa in the final group
\item members - A character vector Names of taxa selected in the final group.
\item abundance - A numeric vector with relative abundance of the final group in each sample.
\item y - Coefficient of determination for the linear regression.
}
}
\description{
This function performs BLS-based Ensemble Quotient Optimization
towards annotation-free microbiome coarse-graining.
}
\details{
The BLS-based algorithm searches for a group of species
which as a whole is most strongly correlated with a continuous functional trait variable.
To run BLS more effciently, it is recommended to reduce the total number of taxa
by removing the rare singletons (e.g., < 300 taxa in the final input) and apply
strong regularization (e.g., Nmax<20).
}
