% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQO_ga.R
\name{EQO_ga}
\alias{EQO_ga}
\title{Ensemble Quotient Optimization with genetic algorithm (GA)}
\usage{
EQO_ga(
  pattern,
  M,
  y,
  pk,
  Nmax,
  amin,
  amax,
  maxIter = 500,
  popSize = 200,
  parallel = TRUE,
  monitor = plot
)
}
\arguments{
\item{pattern}{A character speficifying the type of trait, including \code{"u"} for a uniform trait, \code{"c"} for a continuous trait and \code{"d"} for a discrete trait.}

\item{M}{A matrix of taxa in the microbiome (samples as rows and taxa as columns). Rownames and colnames should be specified.}

\item{y}{A vector or matrix of trait. y is optional when pattern is \code{"u"}. y is a vector when pattern is \code{"c"}. y is a binary matrix when pattern is \code{"d"} whose rows are samples and columns are categories.}

\item{pk}{A binary vector indicating partially known functional group based on a priori knowledge, with 1 for species forced to be included in the targeted group and 0 for the other unknown species. Length of this vector should be equal to the total number of species in the microbiome. EQO will then search on the basis of the provided partial known group without removing the designated species.}

\item{Nmax}{A number (default the total number of species) for regularization, specifying the maximal number of taxa allowed in the final group.}

\item{amin}{A number (default 0) specifying the lower bound of average relative abundance of the final group across all samples (only applied to a uniform trait to avoid trivial solutions).}

\item{amax}{A number (default 1) specifying the upper bound of average relative abundance of the final group across all samples (only applied to a uniform trait to avoid trivial solutions).}

\item{maxIter}{A number (default 500) for GA package, specifying the maximum number of iterations to run before the GA search is halted.}

\item{popSize}{A number (default 200) for GA package, specifying the initial population size for the GA search.}

\item{parallel}{Optional argument for GA package (default TRUE), indicating whether parallel computing is needed. See GA documentation for full details.}

\item{monitor}{Optional argument for GA package (default plot). \code{monitor=plot} indicates that algorithm implementation is visualized as a plot. \code{monitor=gaMonitor} indicates that algorithm implementation is printed as text. \code{monitor=FALSE} indicates the output should be suppressed.}
}
\value{
A list with the following components.
\itemize{
\item fitness - optimized value for the objective function in GA
\item x - A binary numeric vector specifying presence/absence of taxa in the final group
\item members - A character vector Names of taxa selected in the final group
\item abundance - A numeric vector with relative abundance of the final group in each sample.
\item performance - A number of the group performance. It equals to coefficient of variation when the trait is uniform, for which a smaller value implies higher stability. It eqauls to correlation coefficient when the trait is continuous, for which a larger absolute value implies stronger association. It equals to generalized coefficient of determination when the trait is discrete, for which a larger value implies better discrimination across different categories.
}
}
\description{
This function performs GA-based Ensemble Quotient Optimization
towards annotation-free microbiome coarse-graining.
}
